+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "pg_search",
     "pg_sparse",
     "shared",
+    "shared_ext"
 ]
 
 [profile.dev]
