+++ shared/Cargo.toml
@@ -5,18 +5,7 @@ version = "0.1.0"
 edition = "2021"
 license = "AGPL-3.0"
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-doctest = false
-
 [features]
-default = ["pg15"]
-pg12 = ["pgrx/pg12", "pgrx-tests/pg12" ]
-pg13 = ["pgrx/pg13", "pgrx-tests/pg13" ]
-pg14 = ["pgrx/pg14", "pgrx-tests/pg14" ]
-pg15 = ["pgrx/pg15", "pgrx-tests/pg15" ]
-pg16 = ["pgrx/pg16", "pgrx-tests/pg16" ]
-pg_test = []
 telemetry = []
 
 [dependencies]
