+++ shared_ext/src/lib.rs
@@ -1,0 +1,27 @@
+mod logs;
+
+use pgrx::*;
+
+pgrx::pg_module_magic!();
+
+#[allow(clippy::missing_safety_doc)]
+#[allow(non_snake_case)]
+#[pg_guard]
+pub unsafe extern "C" fn _PG_init() {
+    // Initalize Grand Unified Configuration (GUC) variables.
+    shared::gucs::init();
+}
+
+/// This module is required by `cargo pgrx test` invocations.
+/// It must be visible at the root of your extension crate.
+#[cfg(test)]
+pub mod pg_test {
+    pub fn setup(_options: Vec<&str>) {
+        // perform one-off initialization when the pg_test framework starts
+    }
+
+    pub fn postgresql_conf_options() -> Vec<&'static str> {
+        // return any postgresql.conf settings that are required for your tests
+        vec![]
+    }
+}
